Fix nut related issues on the back of the body
One nut recess was not being made as the sides were not touching / shape not closed so the extraction wasn't catching it.
Also increased the depth for the nuts to hopefully be at least flush with the surface after the nut is inserted.

Updated readme to add note about direction for printing.